Installing icu (#1)
Trello: https://trello.com/c/uusDUnPO/642-installing-icu-dependency-for-charlock-holmes

This builds ICU for Charlock Holmes and ensures the configuration is available for Bundler.

The easiest way to create this buildpack was to vendor ICU into the repository and build from scratch. Using apt-get was possible but required about 10 times more code in order to manipulate where the package gets installed.

This downloads and compiles the specified version of Charlock Holmes, making it available under /app/icu for later build packs.